# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MT5 Local Gateway API",
    "description": "MT5 网关：行情 HTTP/SSE + MetaApi 兼容交易 RPC（HTTP JSON）。\n\n**运行模式**\n- `local`（Windows）：`/rpc/*` → 本机 Python `9530`\n- `router`（Linux 统一入口）：`/rpc/*` 按 `X-Account-Id` 转发至 `deploy/accounts.json` 中的 Windows 网关\n\n**行情**：Python bridge → TCP 9627 → Node `/quotes`\n\n**鉴权**（可选）：`Authorization: Bearer <MT5_GATEWAY_API_KEY>` 或 `X-Api-Key`；IP 白名单 `MT5_GATEWAY_IP_WHITELIST`\n\nSwagger: `/docs` | OpenAPI: `/openapi.json`",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://127.0.0.1:9628", "description": "Windows 本机网关 (local)" },
    { "url": "http://43.99.61.45:9628", "description": "Linux 统一入口 (router)" }
  ],
  "tags": [
    { "name": "system", "description": "健康检查" },
    { "name": "quotes", "description": "实时行情" },
    { "name": "trade", "description": "MetaApi 兼容交易。router 模式需 Header `X-Account-Id`（MT5 登录号）" }
  ],
  "paths": {
    "/health": {
      "get": {
        "tags": ["system"],
        "summary": "服务健康检查",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HealthResponse" },
                "example": {
                  "ok": true,
                  "quotes": 3,
                  "trade": true,
                  "mode": "router",
                  "backends": {
                    "60100824": { "ok": true, "gateway": "http://10.0.1.10:9628", "latencyMs": 12, "trade": true }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/quotes": {
      "get": {
        "tags": ["quotes"],
        "summary": "全部品种行情",
        "responses": {
          "200": {
            "description": "行情列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Quote" }
                }
              }
            }
          }
        }
      }
    },
    "/quote/{symbol}": {
      "get": {
        "tags": ["quotes"],
        "summary": "单品种行情",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "example": "XAUUSD" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Quote" }
              }
            }
          },
          "404": { "description": "暂无行情" }
        }
      }
    },
    "/quotes/stream": {
      "get": {
        "tags": ["quotes"],
        "summary": "SSE 行情推送",
        "description": "Content-Type: text/event-stream。事件：snapshot、quote、remove。",
        "responses": {
          "200": { "description": "SSE 流" }
        }
      }
    },
    "/rpc/get_account_information": {
      "get": {
        "tags": ["trade"],
        "summary": "get_account_information",
        "description": "MetaApi 等价：查询 MT 账户信息",
        "parameters": [
          { "$ref": "#/components/parameters/XAccountId" },
          { "$ref": "#/components/parameters/AccountIdQuery" }
        ],
        "security": [{ "GatewayApiKey": [] }, {}],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AccountInfo" }
              }
            }
          },
          "400": { "description": "缺少 X-Account-Id（MT5_REQUIRE_ACCOUNT_ID=1）" },
          "401": { "description": "API Key 无效" },
          "403": { "description": "IP 不在白名单" },
          "404": { "description": "未知 accountId 或账户未配置" },
          "502": { "description": "router 转发 Windows 网关失败" },
          "503": { "description": "交易未启用 (MT5_TRADE_ENABLED=0)" }
        }
      }
    },
    "/rpc/get_positions": {
      "get": {
        "tags": ["trade"],
        "summary": "get_positions",
        "description": "MetaApi 等价：当前全部持仓",
        "parameters": [
          { "$ref": "#/components/parameters/XAccountId" },
          { "$ref": "#/components/parameters/AccountIdQuery" }
        ],
        "security": [{ "GatewayApiKey": [] }, {}],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Position" }
                }
              }
            }
          },
          "400": { "description": "缺少 X-Account-Id" },
          "401": { "description": "API Key 无效" },
          "403": { "description": "IP 不在白名单" },
          "404": { "description": "未知 accountId" },
          "502": { "description": "router 转发失败" }
        }
      }
    },
    "/rpc/get_orders": {
      "get": {
        "tags": ["trade"],
        "summary": "get_orders",
        "description": "MetaApi 等价：当前全部挂单（pending orders）",
        "parameters": [
          { "$ref": "#/components/parameters/XAccountId" },
          { "$ref": "#/components/parameters/AccountIdQuery" }
        ],
        "security": [{ "GatewayApiKey": [] }, {}],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Order" }
                }
              }
            }
          },
          "400": { "description": "缺少 X-Account-Id" },
          "401": { "description": "API Key 无效" },
          "403": { "description": "IP 不在白名单" },
          "404": { "description": "未知 accountId" },
          "502": { "description": "router 转发失败" }
        }
      }
    },
    "/rpc/get_position/{positionId}": {
      "get": {
        "tags": ["trade"],
        "summary": "get_position",
        "parameters": [
          { "$ref": "#/components/parameters/XAccountId" },
          { "$ref": "#/components/parameters/AccountIdQuery" },
          {
            "name": "positionId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "security": [{ "GatewayApiKey": [] }, {}],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Position" }
              }
            }
          },
          "404": { "description": "持仓不存在" }
        }
      }
    },
    "/rpc/get_order/{orderId}": {
      "get": {
        "tags": ["trade"],
        "summary": "get_order",
        "parameters": [
          { "$ref": "#/components/parameters/XAccountId" },
          { "$ref": "#/components/parameters/AccountIdQuery" },
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "security": [{ "GatewayApiKey": [] }, {}],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Order" }
              }
            }
          },
          "404": { "description": "订单不存在" }
        }
      }
    },
    "/rpc/get_history_orders_by_time_range": {
      "get": {
        "tags": ["trade"],
        "summary": "get_history_orders_by_time_range",
        "parameters": [
          { "$ref": "#/components/parameters/XAccountId" },
          { "$ref": "#/components/parameters/AccountIdQuery" },
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "format": "date-time" },
            "example": "2026-06-01T00:00:00Z"
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "format": "date-time" },
            "example": "2026-06-16T23:59:59Z"
          }
        ],
        "security": [{ "GatewayApiKey": [] }, {}],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Order" }
                }
              }
            }
          }
        }
      }
    },
    "/rpc/create_market_buy_order": {
      "post": {
        "tags": ["trade"],
        "summary": "create_market_buy_order",
        "description": "MetaApi 等价：市价开多（立即成交，非挂单）。router 模式必须带 X-Account-Id。",
        "parameters": [
          { "$ref": "#/components/parameters/AccountIdHeader" }
        ],
        "security": [{ "GatewayApiKey": [] }, {}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MarketOrderRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成交成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TradeResult" }
              }
            }
          },
          "400": { "description": "拒单（如 AutoTrading 未开、填充模式错误）" }
        }
      }
    },
    "/rpc/create_market_sell_order": {
      "post": {
        "tags": ["trade"],
        "summary": "create_market_sell_order",
        "description": "MetaApi 等价：市价开空",
        "parameters": [
          { "$ref": "#/components/parameters/AccountIdHeader" }
        ],
        "security": [{ "GatewayApiKey": [] }, {}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MarketOrderRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成交成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TradeResult" }
              }
            }
          }
        }
      }
    },
    "/rpc/close_position": {
      "post": {
        "tags": ["trade"],
        "summary": "close_position",
        "description": "MetaApi 等价：按 positionId 全平",
        "parameters": [
          { "$ref": "#/components/parameters/AccountIdHeader" }
        ],
        "security": [{ "GatewayApiKey": [] }, {}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ClosePositionRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CloseResult" }
              }
            }
          }
        }
      }
    },
    "/rpc/close_position_partially": {
      "post": {
        "tags": ["trade"],
        "summary": "close_position_partially",
        "description": "MetaApi 等价：部分平仓",
        "parameters": [
          { "$ref": "#/components/parameters/AccountIdHeader" }
        ],
        "security": [{ "GatewayApiKey": [] }, {}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ClosePartialRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CloseResult" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "GatewayApiKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "设置 MT5_GATEWAY_API_KEY 时必填；也可用 Header X-Api-Key"
      }
    },
    "parameters": {
      "XAccountId": {
        "name": "X-Account-Id",
        "in": "header",
        "required": false,
        "schema": { "type": "string", "example": "60100824" },
        "description": "MT5 登录号。router 模式用于选择 accounts.json 中的后端网关"
      },
      "AccountIdQuery": {
        "name": "accountId",
        "in": "query",
        "required": false,
        "schema": { "type": "string", "example": "60100824" },
        "description": "GET 类接口可用 query 传递 accountId（优先于 default）"
      }
    },
    "schemas": {
      "HealthResponse": {
        "type": "object",
        "properties": {
          "ok": { "type": "boolean" },
          "quotes": { "type": "integer" },
          "trade": { "type": "boolean" },
          "mode": { "type": "string", "enum": ["local", "router", "off"] },
          "tradeUrl": { "type": "string", "description": "local 模式 Python 地址" },
          "backends": {
            "type": "object",
            "additionalProperties": {
              "type": "object",
              "properties": {
                "ok": { "type": "boolean" },
                "gateway": { "type": "string" },
                "latencyMs": { "type": "integer" },
                "trade": { "type": "boolean" },
                "error": { "type": "string" }
              }
            },
            "description": "router 模式各 accountId 后端探测结果"
          }
        }
      },
      "Quote": {
        "type": "object",
        "properties": {
          "symbol": { "type": "string" },
          "bid": { "type": "number" },
          "ask": { "type": "number" },
          "time": { "type": "integer" },
          "updatedAt": { "type": "integer", "description": "本地更新时间戳 ms" }
        }
      },
      "AccountInfo": {
        "type": "object",
        "properties": {
          "login": { "type": "string" },
          "server": { "type": "string" },
          "balance": { "type": "number" },
          "equity": { "type": "number" },
          "margin": { "type": "number" },
          "freeMargin": { "type": "number" },
          "marginLevel": { "type": "number" },
          "tradeAllowed": { "type": "boolean" }
        }
      },
      "Position": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "positionId": { "type": "string" },
          "symbol": { "type": "string" },
          "volume": { "type": "number" },
          "type": { "type": "string", "example": "POSITION_TYPE_BUY" },
          "side": { "type": "string", "enum": ["BUY", "SELL"] },
          "comment": { "type": "string" },
          "brokerComment": { "type": "string" },
          "openPrice": { "type": "number" },
          "currentPrice": { "type": "number" },
          "profit": { "type": "number" }
        }
      },
      "Order": {
        "type": "object",
        "properties": {
          "orderId": { "type": "string" },
          "symbol": { "type": "string" },
          "volume": { "type": "number" },
          "comment": { "type": "string" }
        }
      },
      "MarketOrderRequest": {
        "type": "object",
        "required": ["symbol", "volume"],
        "properties": {
          "symbol": { "type": "string", "example": "XAUUSD" },
          "volume": { "type": "number", "example": 0.01 },
          "stopLoss": { "type": "number" },
          "takeProfit": { "type": "number" },
          "comment": { "type": "string", "example": "order-no-123" },
          "clientId": { "type": "string", "description": "可选，与 MetaApi tradeOptions 一致；本地仅记录不转发" }
        }
      },
      "ClosePositionRequest": {
        "type": "object",
        "required": ["positionId"],
        "properties": {
          "positionId": { "type": "string" },
          "comment": { "type": "string" },
          "clientId": { "type": "string" }
        }
      },
      "ClosePartialRequest": {
        "allOf": [
          { "$ref": "#/components/schemas/ClosePositionRequest" },
          {
            "type": "object",
            "required": ["volume"],
            "properties": {
              "volume": { "type": "number", "example": 0.01 }
            }
          }
        ]
      },
      "TradeResult": {
        "type": "object",
        "properties": {
          "orderId": { "type": "string" },
          "positionId": { "type": "string" },
          "symbol": { "type": "string" },
          "actionType": { "type": "string", "example": "ORDER_TYPE_BUY" },
          "volume": { "type": "number" },
          "status": { "type": "string", "example": "success" },
          "error": { "nullable": true }
        }
      },
      "CloseResult": {
        "type": "object",
        "properties": {
          "orderId": { "type": "string" },
          "positionId": { "type": "string" },
          "status": { "type": "string" },
          "error": { "nullable": true }
        }
      }
    }
  }
}